pyproject.toml
[project]
name = "fintech-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai==0.8.5",
    "python-dotenv==1.1.1",
    "pydantic==2.11.9",
    "pydantic-settings==2.11.0",
    "fastapi== 0.117.1",
    "uvicorn[standard]==0.37.0",
    "sentence-transformers==5.1.1",
    "faiss-cpu==1.12.0",
    "numpy==2.3.3",
    "loguru==0.7.3",
]
